Containerized scripts respect VERBOSE env-var
Previously, the VERBOSE environment variable was used only to control
the verbosity of the build process. This commit propagates the setting
of VERBOSE down to containerized scripts, allowing them to output
verbosely if VERBOSE is set on the container host.

This is done by  sed-ing the string "__VERBOSE" into either the empty
string or "-v" in the Dockerfiles of scripts, depending on whether the
environment variable VERBOSE has been set in the build environment.