Output stored in data container
Output from all experiments will now be stored in a data container,
rather than in a local directory.

This commit makes the necessary changes to the Makefile, and also
creates a test container. The test container accesses data from the data
container in order to run tests. There are currently no tests.