Deps generation takes base packs into account
This commit makes the deps_to_ninja container not generate build targets
for packages in the base and base-devel groups.  It does this by taking
into account the results of the generate_base_package_names container.

This commit also includes tests to ensure that the generated ninja file
does not contain any build targets in the base or base-devel groups.