Output stored in data container
Output from containers is no longer stored locally, but rather written
to a data container. This is to facilitate sharing of data between
containers, and running tests on container output.

A test container is also committed. A test suite called test/foo.py is
intended to test the output of container foo.