Dependency file generation & using ninja for build
Summary:
--------

Tuscan now outputs a `build.ninja' file which describes how to build
every Arch Linux package. This is enabled by several changes:

* The introduction of the get_base_package_names experiment;
* Refactoring deps_to_ninja so that shared utilities are moved out;
* Using Ninja for the overall build rather than a Makefile.

Details:
--------

The get_base_package_names experiment returns the `bootstrap set' of
packages. That is, the set of packages that must be present as binaries
in order to build the rest of Arch Linux. This is necessary to provide a
starting point for the build. get_base_package_names and the changes in
provides.py also resolve issues with circular dependencies and package
name provision; more details are in the README.

Common utilities that shall be used by all experiments have been moved
into utilities.py. This will ensure that experiments are all using the
same conventions, paths to write data, etc. and eliminate code
duplication.

This commit also switches to Ninja for the overall build. The file
tuscan.py generates a ninja file and runs it in order to run the
experiments. Dependency information between experiments and data
containers is not contained in tuscan.py, but rather in human-writable
YAML files in each experiment's directory. This should help to isolate
changes to the build, i.e. new dependency and experiment information
goes in YAML files, whereas different build logic should cause a change
in tuscan.py.