Adding toolchain directory
This commit introduces the toolchain directory, which contains
configuration options for each toolchain that we wish to build with.

Each directory contains at least a Dockerfile and makepkg.conf. The
Dockerfile sets up an Arch Linux container with the toolchain, e.g. by
downloading and installing the toolchain. The makepkg.conf is installed
in the default location on Arch (/etc/makepkg.conf) and can be used to
override makepkg's default settings.