Adding custom repository builder
This commit introduces the custom_repository analysis, which builds a
local repository of packages in the data container. It also initialises
this repository with `bootstrap set' packages, that is, packages that
need to be present as binaries in order to build all other packages.

This commit also changes logic in tuscan.py. We do not want the
repository container to be shared amongst toolchains, since packages
built with one toolchain should not be installed when compiling with
another toolchain. Thus, the name of the data container is now
parametric over the name of the toolchain; the string __TOOLCHAIN__ gets
replaced with the name of the toolchain in yaml files.