Hybrid package creation
This commit enables the creation of hybrid packages---packages
containing both vanilla and toolchain-built data.

New stages
----------

The create_toolchain_repo stage is added. This creates an empty
package repository in a data container. This package repository is used
by the make_package stage:
  * when a package is successfully built, it is added to the package
    repository;
  * dependencies of a build are installed from the package repository
    before the build begins.

The install_bootstrap stage is added. This installs all vanilla packages
that are necessary to build all other packages (the bootstrap packages).
This container is committed after it finishes running, so that
containers running the make_package stage can use it as a base image.
This stage is run just once, since installing all the bootstrap packages
is a time-consuming process.

Local Arch Mirror
-----------------

Packages are now synced from a local arch mirror, rather than upstream.
This saves bandwidth and is more polite, as well as making the
experiment reproducible.

Hybrid package creation
-----------------------

The make_package and create_toolchain_repo stages now create hybrid
packages. When a package is built, the hybrid packages that it depends
on are installed, and the compiler uses toolchain-built libraries and
headers.

Logging
-------

The make_package, install_bootstrap and create_toolchain_repo stages
emits JSON-structured logs that can be parsed into HTML reports.

Bugfix: file exists in filesystem

Before this commit, the toolchain portion of hybrid packages would be
installed in a global location and not uninstalled. However, the package
database is local to each container. This means that containers
attempting to install a dependency whose files already existed on the
file system would fail.

This commit makes the installation location of toolchain files local to
each container, while keeping the toolchain archive global.

Compiler wrapper
----------------

This commit introduces a rudimentary compiler wrapper that replaces the
native GCC, ld and ar programs. It prints out the name of the program it
is replacing and then returns non-zero.

This is so that we can detect when toolchain builds are ignoring our
setting of CC and hardcoding the name of the native compiler.