make_package logs info on provided packages
The make_package stage now writes a list of packages that are provided
by its build to the build log. This is done so that we know what
packages were supposed to have been built by this build.

Any future build that depends on one of the packages will know which
build to blame if it cannot install some of its dependencies. This will
allow us to show a graph of what builds are causing lots of future
builds to fail.

This commit has necessitated refactoring out some common code to
utilities.py.