Missing source is detected earlier in build
This commit adds a check for source directories that don't exist on the
permanent source volume at all (as opposed to source that can't be
obtained for some other reason). It is an earlier check for missing
source than the checks we already had, and fails the build with a log
message rather than causing a crash.