Bear now logs return codes of execed processes
There are now two kinds of record generated from the bear tool: those
emitted when a process is execed, and those emitted when a process
exits.

The records are discriminated by the new "kind" field, which shall be
set to either "EXEC" or "EXIT". "EXIT" records contain the return code,
PID, and parent PID of the exiting process.