Allow user to specify toolchain to post-process
The user can now optionally specify a list of toolchains to
post-process, instead of post-processing the results of all toolchains
(which remains the default). This is because post-processing it taking a
large amount of main memory, and running the post-processing one
toolchain at a time can help the machine to not thrash.