More accurate build count on HTML summary pages
Summary pages can display the total number of builds on that page.
Previous to this commit, summary pages that displayed builds for more
than one toolchain would count each build once for each toolchain, i.e.
tripling the apparent number of builds if there are three toolchains on
display.

This commit changes that behaviour so that the "total" number of builds
is the number of builds divided by the number of toolchains on display.