Improved memory efficiency of blocker post-process
Post-processed results are no longer stored in an inter-process data
structure while waiting to have their blockers calculated, as this was a
huge waste of memory. Instead, results are written to disk after all
post-processing except for blocker analysis is done. When all results
have been written, only the necessary data for blocker analysis is re-
loaded from disk; the on-disk results are then updated with the blocker
results.

Due to this improvement, the blocker analysis has been re-enabled by
default.