Toolchains are created once and saved locally
Toolchains are now downloaded and built into a locally-mounted
directory under the directory `sysroots`, rather than a data container.
If a particular toolchain already exists in `sysroots`, then it will not
be re-downloaded on subsequent invocations of `tuscan build` for that
toolchain.

This is to facilitate reproducibility of results. Before this commit,
there was no provision to load a pre-created toolchain into tuscan; each
toolchain would be downloaded from the upstream every time tuscan was
run. This commit will allow us to distribute pre-created toolchains to
others; tuscan will then use those toolchains rather than downloading
the latest version.