revert musl to clang 3.8.1
This commit fixes some bitrot that set in since August. The musl
toolchain is now working, albeit the toolchain itself is not
bootstrapped. Binutils and LLVM are built with the host compiler, and
musl is built with that LLVM. We're also adding /sysroot/lib to the
list dynamic lookup paths in /etc/ld.so.conf.